requirements.txt
alembic==1.13.1
asgiref==3.8.1
async-timeout==4.0.3
blinker==1.7.0
boto3==1.34.72
botocore==1.34.79
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
Flask-Cors==4.0.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.0.3
heimdall-tools==1.2.2
idna==3.6
importlib-metadata==7.1.0
importlib-resources==6.4.0
influxdb-client==1.43.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mysql-connector-python==8.3.0
packaging==24.0
paho-mqtt==1.6.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
redis==5.0.4
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
typing-extensions==4.11.0
urllib3==1.26.18
werkzeug==3.0.2
zipp==3.18.1
hvac==2.3.0